requirements.txt
# Flask template
# --------------
alembic>=0.3.6
dealer>=0.1.8
Flask>=0.9
Flask-Admin>=1.0.2
Flask-Babel>=0.8
Flask-Bootstrap>=2.1.0-1
Flask-Cache>=0.8.0
Flask-Collect>=0.1.3
Flask-DebugToolbar>=0.7.1
Flask-Login>=0.1.3
Flask-Mail>=0.7.3
Flask-Mixer>=0.2.1
Flask-Principal>=0.3.3
Flask-Rauth>=0.2.2
Flask-Script>=0.5.0
Flask-SQLAlchemy>=0.16
Flask-Testing>=0.4
mock>=1.0b1
straight.plugin>=1.4.0-post-1
